Pushed v1.16 code
- The mouse system has been rewritten so that we don't need to do mouse
  capturing while interacting with GUI widgets, which could be buggy sometimes.
- The different Disk Op. item paths (mod./instr./smp./pat./trk.) now behave
  exactly like real FT2. I.e. if they haven't been initialized before, they
  will be set to the current working directory.
- Scrollbar thumbs are now limited to minimum 9 pixels in length/height, to
  prevent them from being difficult to click on.

+++ src/ft2_sample_ed.c
@@ -1440,7 +1440,7 @@ void rangeAll(void)
 	smpEd_Rx2 = smpEd_ScrPos + smpEd_ViewSize;
 }
 
-static void zoomSampleDataIn(int32_t step, int16_t x)
+static void zoomSampleDataIn(int32_t step, int32_t x)
 {
 	int32_t tmp32, minViewSize;
 	int64_t newScrPos64;
@@ -1482,7 +1482,7 @@ static void zoomSampleDataIn(int32_t step, int16_t x)
 	updateScrPos();
 }
 
-static void zoomSampleDataOut(int32_t step, int16_t x)
+static void zoomSampleDataOut(int32_t step, int32_t x)
 {
 	int32_t tmp32;
 	int64_t newViewSize64;
